Conflicts:
	packages.txt

+++ install.sh
@@ -121,13 +121,21 @@ for rc in vim gvim; do
         touch $HOME/.${rc}local
     fi
 done
-link $basedir/.vim/_vimoutliner $HOME/.vimoutliner
-link $basedir/.vim/_vimoutlinerrc $HOME/.vimoutlinerrc
+
+# create vim_local here, because conditionals in vim are hard.
+if [ ! -e $HOME/.vim_local ]; then
+  touch $HOME/.vim_local
+fi
 
 note "Initializing tools..."
 if has git; then
-    # Post-install scripts might customize this further.
+  # Post-install scripts might customize this further.
+  if [ ! -e $HOME/.gitconfig ]; then
     cp -v $basedir/.gitconfig.base $HOME/.gitconfig
+    if [ -e $HOME/.gitconfig.local ]; then
+      cat $HOME/.gitconfig.local >> $HOME/.gitconfig
+    fi
+  fi
 fi
 
 note "Running post-install script, if any..."
